# gateway/openapi.yaml
openapi: 3.0.3
info:
  title: kkode Gateway API
  version: 0.1.0
  description: |
    kkode의 session-aware coding agent gateway API예요.
    REST는 session/run 제어면을 담당하고, SSE는 event replay/stream을 담당해요.
servers:
  - url: http://127.0.0.1:41234
security:
  - bearerAuth: []
paths:
...
  /api/v1/mcp/servers:
    get:
      summary: MCP server manifest 목록을 반환해요.
      parameters:
        - $ref: '#/components/parameters/Enabled'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: MCP server manifest 목록이에요.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceListResponse'
    post:
      summary: MCP server manifest를 저장해요.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Resource'
      responses:
        '201':
          description: 저장된 MCP server manifest예요.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resource'
  /api/v1/mcp/servers/{resource_id}:
    get:
      summary: MCP server manifest를 조회해요.
      parameters:
        - $ref: '#/components/parameters/ResourceID'
      responses:
        '200':
          description: MCP server manifest예요.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resource'
        '404':
          $ref: '#/components/responses/Error'
    put:
      summary: MCP server manifest를 갱신해요.
      parameters:
        - $ref: '#/components/parameters/ResourceID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Resource'
      responses:
        '200':
          description: 갱신된 MCP server manifest예요.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resource'
    delete:
      summary: MCP server manifest를 삭제해요.
      parameters:
        - $ref: '#/components/parameters/ResourceID'
      responses:
        '204':
          description: 삭제됐어요.
        '404':
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
...
    ResourceID:
      in: path
      name: resource_id
      required: true
      schema:
        type: string
    Enabled:
      in: query
      name: enabled
      schema:
        type: boolean
    Limit:
      in: query
      name: limit
      schema:
        type: integer
        default: 100
        maximum: 500
  responses:
    Error:
      description: 표준 error envelope이에요.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
...
    ResourceListResponse:
      type: object
      required: [resources]
      properties:
        resources:
          type: array
          items:
            $ref: '#/components/schemas/Resource'
    Resource:
      type: object
      required: [name]
      properties:
        id:
          type: string
        kind:
          type: string
          enum: [mcp_server, skill, subagent]
        name:
          type: string
        description:
          type: string
        enabled:
          type: boolean
          default: true
        config:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string
            details:
              type: object
              additionalProperties: true